Refactor: extract MediaChangeDispatcher for testability
Move the onMediaChanged retry logic from OverlayService into a new
MediaChangeDispatcher class. All Android-framework side-effects are
accessed through constructor lambdas (queryLatestMedia, handler) so
the class can be exercised in plain JVM unit tests without Robolectric.

OverlayService wires MediaChangeDispatcher in onCreate() and delegates
the ContentObserver.onChange callback to it — no behavior change.
